made SignUrlExpires type 'long'

+++ src/Idfy.Signature.Client/Idfy.Signature.Models/Signers/Signer.cs
@@ -55,7 +55,7 @@ namespace Idfy.Signature.Models.Signers
         /// <summary>
         /// How long before the signers url should expire? Utc date in ticks. This can be set if you only want a limited time to live for each sign url (If you generate a new url at a later time this will also have this limited lifetime). Defaults to the document lifetime.
         /// </summary>
-        public int SignUrlExpires { get; set; }
+        public long SignUrlExpires { get; set; }
     }
 
     public class UI
